Update README.md
Imagine you are reading an interesting topic from the internet. Suppose the topic is divided into 100 subtopics where each topic begins on a different page that makes a topic of 100 different pages. Now imagine you want to download all the topics for offline reading. Of course you can navigate to each page on your web browser and download the pages manually one page at a time. But that would be boring right? plus it will take a good amount of time. This project shows how we can automate the whole process and download all the pages with just a few lines of python code. 
For this project, I made use of the request library, bs4 library and a html parser(lxml)
So before you run this code on your device, make sure these libraries are already installed either in a virtual environment or in your local machine.